sclang: ensure minimum stack size
the sclang interpreter does not do any stack overflow checks. the stack
size of sclang threads is defaulted to 64, which is quite low. especially
it caused a gc corruption when running certain code in routines.

this patch increases the default stack size and adds a lower bound to the
stack size

Reported-by: Jonatan Liljedahl <lijon@kymatica.com>
Signed-off-by: Tim Blechmann <tim@klingt.org>