move sclang to the MacOS directory with the app executable
QT plugin loading is determined in part by the qt.conf file, which points to a Plugin dir. sclang seems to get confused by being adjacent to that file, and misunderstanding where it should load plugins from. Simply moving sclang is the simplest way I found to ensure both the IDE and sclang load plugins and do runtime linking successfully. There MAY be a better solution to be had - enabling the QT_DEBUG_PLUGINS env flag helps show what's going on - but this is working, and /MacOS is a fine and possibly more appropriate place for sclang anyway, so there it stays.